Refs #44: Remove about and help menu action classes, clean up control panel and cyactivator class
